pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dep-age"
version = "1.0.0"
description = "Cross-language dependency age analyzer — scan lock files for staleness, CVEs, and update urgency"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "dep-age contributors" }]
keywords = ["dependencies", "security", "audit", "outdated", "cve", "lockfile"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.9,<1.0",
    "rich>=13.0,<14.0",
    "httpx>=0.27,<1.0",
    "python-dateutil>=2.8,<3.0",
    "diskcache>=5.6.3,<6.0",
    "pyyaml>=6.0,<7.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<10.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/dep-age/dep-age"
Repository = "https://github.com/dep-age/dep-age"
Issues = "https://github.com/dep-age/dep-age/issues"

[project.scripts]
dep-age = "dep_age.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/dep_age"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "S"]
ignore = ["S101"]

[tool.coverage.run]
source = ["dep_age"]

[tool.coverage.report]
fail_under = 95
show_missing = true
